test : 좋아요 테스트코드 추가

+++ src/main/kotlin/io/csbroker/apiserver/service/post/PostServiceImpl.kt
@@ -81,7 +81,7 @@ class PostServiceImpl(
         user: User?,
     ) = PostResponseDto(
         post,
-        likeCount = postLikes.count(),
+        likeCount = postLikes.count().toLong(),
         isLiked = postLikes.any { like -> like.user == user },
         comments = comments.map { comment ->
             CommentResponseDto(
